refactor: change structure and rename the svg template
