+++ packages/vckit-oa-renderers/examples/templates/samples/index.ts
@@ -1,3 +1,3 @@
 export * from "./customTemplateSample";
 export * from "./qrcodeTemplateSample";
-export * from "./svgTemplateSample";
+export * from "./purchaseOrderSample";
