--- packages/vckit-oa-renderers/examples/templates/svgRenderer/SvgRenderer.stories.mdx
@@ -1,22 +1,0 @@
-import { Meta, Preview, Props, Description, Story } from "@storybook/addon-docs/blocks";
-import { object } from "@storybook/addon-knobs";
-import SVGRenderer from "./svgRenderer";
-import { svgTemplateCertificate } from "../samples";
-
-<Meta title="MDX|SVGRenderer" component={SVGRenderer} />
-
-# SVGRenderer component
-
-<Description of={SVGRenderer} />
-
-# Props
-
-<Props of={SVGRenderer} />
-
-# Usage
-
-<Preview>
-  <Story name="basic sample">
-    <SVGRenderer document={object("document", svgTemplateCertificate)} />
-  </Story>
-</Preview>
