--- packages/vckit-oa-renderers/examples/templates/svgRenderer/index.tsx
@@ -1,9 +1,0 @@
-import SVGRenderer from "./svgRenderer";
-
-export const templates = [
-  {
-    id: "svgRenderer",
-    label: "SVG Renderer",
-    template: SVGRenderer
-  }
-];
