+++ packages/vckit-oa-renderers/examples/templates/svgTemplate/index.tsx
@@ -1,0 +1,9 @@
+import SvgTemplate from "./svgTemplate";
+
+export const templates = [
+  {
+    id: "svg",
+    label: "SVG Template",
+    template: SvgTemplate
+  }
+];
