+++ packages/vckit-oa-renderers/src/templates/samples/purchaseOrderSample.ts
@@ -1,47 +1,7 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>iFrame Rendering</title>
-  </head>
-  <script src="https://unpkg.com/penpal@5.3.0/dist/penpal.min.js"></script>
-  <body>
-    <div>
-      <button onclick="renderDocument()">
-        Render OA document
-      </button>
-      <button onclick="renderSVGDocument()">
-        Render SVG document
-      </button>
-    </div>
-    <div id="template-selector"></div>
-    <iframe id="iframe" src="http://localhost:3000" style="width: 100%; border: 0px; overflow: hidden;"></iframe>
-  </body>
-
-  <script type="text/javascript">
-    const iframe = document.getElementById("iframe");
-    const templateSelector = document.getElementById("template-selector");
-    let frame;
-
-    const oaDocument = {
-      name: "John Doe",
-      institute: "Institute of John Doe",
-      issuers: [
-        {
-          name: "institute of blockchain"
-        }
-      ],
-      $template: {
-        name: "custom",
-        type: "EMBEDDED_RENDERER",
-        url: "http://localhost:3000"
-      },
-      foo: {
-        title: "Bar is awesome"
-      }
-    };
-
-    const SVGDocument = {
-      svg: `<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+import { v2 } from "@govtechsg/open-attestation";
+const rendererApp = process.env.RENDERER_APP || "http://localhost:3000";
+// The svg is copied from ./purchaseOrderForm.svg
+const svg = `<?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <!-- Created with Inkscape (http://www.inkscape.org/) -->
 
 <svg
@@ -1184,190 +1144,166 @@ gg==
       <tspan id="tspan702-6-4-7-7" x="702.41644" y="307.1767" style="fill:#000000">{{items.1.lineItemTotalPrice.price}} {{items.1.lineItemTotalPrice.priceCurrency}}</tspan>
     </text>
   </g>
-</svg>`,
-      issuers: [
-        {
-          name: "institute of blockchain"
+</svg>`;
+
+interface PersonalInformation {
+  type: string[];
+  location: {
+    type: string[];
+    address: {
+      type: string[];
+      name: string;
+      streetAddress: string;
+      addressLocality: string;
+      addressRegion: string;
+      postalCode: string;
+      addressCountry: string;
+    };
+  };
+}
+
+interface Item {
+  type: string[];
+  product: {
+    type: string[];
+    id: string;
+    description: string;
+    weight: {
+      type: string[];
+      unitCode: string;
+      value: string;
+    };
+  };
+  itemCount: number;
+  grossWeight: {
+    type: string[];
+    value: string;
+    unitCode: string;
+  };
+  lineItemTotalPrice: {
+    type: string[];
+    price: number;
+    priceCurrency: string;
+  };
+}
+
+export interface PurchaseOrderSample extends v2.OpenAttestationDocument {
+  svg: string;
+  type: string[];
+  purchaseOrderNo: string;
+  orderDate: string;
+  buyer: PersonalInformation;
+  seller: PersonalInformation;
+  items: Item[];
+  totalWeight: object;
+  totalOrderAmount: object;
+  $template: v2.TemplateObject;
+}
+
+export const purchaseOrderSample: PurchaseOrderSample = {
+  svg,
+  issuers: [
+    {
+      name: "institute of blockchain"
+    }
+  ],
+  $template: {
+    name: "svg",
+    type: v2.TemplateType.EmbeddedRenderer,
+    url: rendererApp
+  },
+  type: ["PurchaseOrder"],
+  purchaseOrderNo: "fe71665a-e7b3-49ba-ac89-82fc2bf1e877",
+  orderDate: "2021-02-21",
+  buyer: {
+    type: ["Organization"],
+    location: {
+      type: ["Place"],
+      address: {
+        type: ["PostalAddress"],
+        name: "Generic Motors of America",
+        streetAddress: "12 Generic Motors Dr",
+        addressLocality: "Detroit",
+        addressRegion: "Michigain",
+        postalCode: "48232-5170",
+        addressCountry: "USA"
+      }
+    }
+  },
+  seller: {
+    type: ["Organization"],
+    location: {
+      type: ["Place"],
+      address: {
+        type: ["PostalAddress"],
+        name: "Aishi Metal Shinzo Co., Ltd.",
+        streetAddress: "1651, Shimonakano, Yoshida",
+        addressLocality: "Tsubame-shi",
+        addressRegion: "Niigata-ken",
+        postalCode: "959-0215",
+        addressCountry: "Japan"
+      }
+    }
+  },
+  items: [
+    {
+      type: ["TradeLineItem"],
+      product: {
+        type: ["Product"],
+        id: "https://aishi-metal-shinzo.example.com/products/UNS-S30400-chromium-nickel-stainless-steel-roll",
+        description: "UNS S30400 chromium-nickel stainless steel roll",
+        weight: {
+          type: ["QuantitativeValue"],
+          unitCode: "lbs",
+          value: "16500"
         }
-      ],
-      $template: {
-        name: "svg",
-        type: "EMBEDDED_RENDERER",
-        url: "http://localhost:3000"
       },
-      type: ["PurchaseOrder"],
-      purchaseOrderNo: "fe71665a-e7b3-49ba-ac89-82fc2bf1e877",
-      orderDate: "2021-02-21",
-      buyer: {
-        type: ["Organization"],
-        location: {
-          type: ["Place"],
-          address: {
-            type: ["PostalAddress"],
-            name: "Generic Motors of America",
-            streetAddress: "12 Generic Motors Dr",
-            addressLocality: "Detroit",
-            addressRegion: "Michigain",
-            postalCode: "48232-5170",
-            addressCountry: "USA"
-          }
-        }
-      },
-      seller: {
-        type: ["Organization"],
-        location: {
-          type: ["Place"],
-          address: {
-            type: ["PostalAddress"],
-            name: "Aishi Metal Shinzo Co., Ltd.",
-            streetAddress: "1651, Shimonakano, Yoshida",
-            addressLocality: "Tsubame-shi",
-            addressRegion: "Niigata-ken",
-            postalCode: "959-0215",
-            addressCountry: "Japan"
-          }
-        }
-      },
-      items: [
-        {
-          type: ["TradeLineItem"],
-          product: {
-            type: ["Product"],
-            id: "https://aishi-metal-shinzo.example.com/products/UNS-S30400-chromium-nickel-stainless-steel-roll",
-            description: "UNS S30400 chromium-nickel stainless steel roll",
-            weight: {
-              type: ["QuantitativeValue"],
-              unitCode: "lbs",
-              value: "16500"
-            }
-          },
-          itemCount: 5,
-          grossWeight: {
-            type: ["QuantitativeValue"],
-            value: "82500",
-            unitCode: "lbs"
-          },
-          lineItemTotalPrice: {
-            type: ["PriceSpecification"],
-            price: 5200,
-            priceCurrency: "USD"
-          }
-        },
-        {
-          type: ["TradeLineItem"],
-          product: {
-            type: ["Product"],
-            id:
-              "https://aishi-metal-shinzo.example.com/products/Galvannealed-ASTM-A-653-zinc-iron-alloy-coated-steel-sheet",
-            description: "Galvannealed ASTM A-653 zinc-iron alloy-coated steel sheet",
-            weight: {
-              type: ["QuantitativeValue"],
-              value: "12680",
-              unitCode: "lbs"
-            }
-          },
-          itemCount: 20,
-          grossWeight: {
-            type: ["QuantitativeValue"],
-            value: "253600",
-            unitCode: "lbs"
-          },
-          lineItemTotalPrice: {
-            type: ["PriceSpecification"],
-            price: 4400,
-            priceCurrency: "USD"
-          }
-        }
-      ],
-      totalWeight: {
+      itemCount: 5,
+      grossWeight: {
         type: ["QuantitativeValue"],
-        value: "336100",
+        value: "82500",
         unitCode: "lbs"
       },
-      totalOrderAmount: {
+      lineItemTotalPrice: {
         type: ["PriceSpecification"],
-        price: 9600,
+        price: 5200,
         priceCurrency: "USD"
       }
-    };
-
-    const dispatch = action => {
-      const { type, payload } = action;
-      if (type === "UPDATE_HEIGHT") {
-        iframe.height = payload;
+    },
+    {
+      type: ["TradeLineItem"],
+      product: {
+        type: ["Product"],
+        id:
+          "https://aishi-metal-shinzo.example.com/products/Galvannealed-ASTM-A-653-zinc-iron-alloy-coated-steel-sheet",
+        description: "Galvannealed ASTM A-653 zinc-iron alloy-coated steel sheet",
+        weight: {
+          type: ["QuantitativeValue"],
+          value: "12680",
+          unitCode: "lbs"
+        }
+      },
+      itemCount: 20,
+      grossWeight: {
+        type: ["QuantitativeValue"],
+        value: "253600",
+        unitCode: "lbs"
+      },
+      lineItemTotalPrice: {
+        type: ["PriceSpecification"],
+        price: 4400,
+        priceCurrency: "USD"
       }
-    };
-
-    const connection = Penpal.connectToChild({
-      iframe,
-      methods: {
-        dispatch
-      }
-    });
-
-    const renderDocumentAction = {
-      type: "RENDER_DOCUMENT",
-      payload: {
-        document: oaDocument
-      }
-    };
-
-    const renderSVGDocumentAction = {
-      type: "RENDER_DOCUMENT",
-      payload: {
-        document: SVGDocument
-      }
-    };
-
-    const getTemplatesAction = {
-      type: "GET_TEMPLATES",
-      payload: oaDocument
-    };
-
-    const getSVGTemplatesAction = {
-      type: "GET_TEMPLATES",
-      payload: SVGDocument
-    };
-
-    const renderDocument = () => {
-      if (!frame) console.warn("frame connection not established.");
-      frame.dispatch(renderDocumentAction);
-    };
-
-    const renderSVGDocument = () => {
-      if (!frame) console.warn("frame connection not established.");
-      frame.dispatch(renderSVGDocumentAction);
-    };
-
-    const buildTemplateButtons = templates => {
-      templates.forEach((tab, index) => {
-        const btn = document.createElement("button");
-        btn.innerHTML = tab.label;
-        btn.id = `selector-${tab.id}`;
-        btn.onclick = () => {
-          frame.dispatch({
-            type: "SELECT_TEMPLATE",
-            payload: tab.id
-          });
-        };
-        templateSelector.appendChild(btn);
-      });
-    };
-
-    const renderTemplateSelector = () => {
-      if (!frame) console.warn("frame connection not established.");
-      frame.dispatch(getTemplatesAction).then(templates => {
-        buildTemplateButtons(templates);
-      });
-
-      frame.dispatch(getSVGTemplatesAction).then(templates => {
-        buildTemplateButtons(templates);
-      });
-    };
-
-    connection.promise.then(child => {
-      frame = child;
-      renderTemplateSelector();
-    });
-  </script>
-</html>
+    }
+  ],
+  totalWeight: {
+    type: ["QuantitativeValue"],
+    value: "336100",
+    unitCode: "lbs"
+  },
+  totalOrderAmount: {
+    type: ["PriceSpecification"],
+    price: 9600,
+    priceCurrency: "USD"
+  }
+};
